+++ mathics/builtin/datentime.py
@@ -25,6 +25,7 @@ from mathics.core.list import ListExpression
 from mathics.core.symbols import Symbol, SymbolList
 from mathics.core.systemsymbols import (
     SymbolAborted,
+    SymbolFailed,
     SymbolInfinity,
     SymbolNull,
     SymbolRowBox,
@@ -1040,14 +1041,21 @@ class Pause(Builtin):
 
     def apply(self, n, evaluation):
         "Pause[n_]"
-        sleeptime = n.to_python()
-        if not isinstance(sleeptime, (int, float)) or sleeptime < 0:
+        sleeptime = float(n.to_python())
+        if not isinstance(sleeptime, float) or sleeptime < 0:
             evaluation.message(
                 "Pause", "numnm", Expression(SymbolPause, from_python(n))
             )
             return
-
-        time.sleep(sleeptime)
+        # This checks each 10ms if the evaluation
+        # was stopped.
+        while sleeptime > 0.01:
+            sleeptime = sleeptime - 0.01
+            time.sleep(0.01)
+            if evaluation.timeout:
+                return SymbolNull
+        if sleeptime > 0:
+            time.sleep(sleeptime)
         return SymbolNull
 
 
@@ -1088,7 +1096,7 @@ class Now(Predefined):
         return Expression(SymbolDateObject.evaluate(evaluation))
 
 
-if sys.platform != "win32" and not hasattr(sys, "pyston_version_info"):
+if sys.platform != "win32":
 
     class TimeConstrained(Builtin):
         r"""
@@ -1105,24 +1113,24 @@ if sys.platform != "win32" and not hasattr(sys, "pyston_version_info"):
         the evaluation continues after the timeout. However, at the end of the evaluation, the function will return '$Aborted' and the results will not affect
         the state of the \Mathics kernel.
 
-        """
 
         # FIXME: these tests sometimes cause SEGVs which probably means
         # that TimeConstraint has bugs.
 
         # Consider testing via unit tests.
-        # >> TimeConstrained[Integrate[Sin[x]^1000000,x],1]
-        # = $Aborted
+        >> TimeConstrained[Pause[5];x,.1]
+         = $Aborted
 
-        # >> TimeConstrained[Integrate[Sin[x]^1000000,x], 1, Integrate[Cos[x],x]]
-        # = Sin[x]
+        >> TimeConstrained[Pause[5];Integrate[Sin[x],x], .1, Integrate[Cos[x],x]]
+         = Sin[x]
 
-        # >> s=TimeConstrained[Integrate[Sin[x] ^ 3, x], a]
-        #  : Number of seconds a is not a positive machine-sized number or Infinity.
-        #  = TimeConstrained[Integrate[Sin[x] ^ 3, x], a]
+        >> s=TimeConstrained[Integrate[Sin[x] ^ 3, x], a]
+         : Number of seconds a is not a positive machine-sized number or Infinity.
+         = TimeConstrained[Integrate[Sin[x] ^ 3, x], a]
 
-        # >> a=1; s
-        # =  Cos[x] (-5 + Cos[2 x]) / 6
+        >> a=5.; s
+         = Cos[x] (-3 + Cos[x] ^ 2) / 3
+        """
 
         attributes = hold_all | protected
         messages = {